#  Batch suggest in ensemble backends