This adds the `_suggest_batch` method to ensemble backends. 

In all ensemble backends (simple, PAV, NN) the suggestions from the source projects are fetched by using batched suggest calls (this is on project level, so whether a backend actually uses batched suggest depends on the backend).

In case of NN ensemble the prediction (now via `_merge_hit_sets_from_sources`) is performed on the whole batch of the base suggestions in one call of the NN model. The prediction is made using the model's `__call__()` method instead of `predict()` as it is the [recommended way for "small numbers of inputs that fit in one batch"](https://www.tensorflow.org/api_docs/python/tf/keras/Model#predict) and it should offer better performance; this should also fix #674.

Also the [EnsembleOptimizer is made to use batched suggest](https://github.com/NatLibFi/Annif/pull/677/commits/5e392564edd179355d61039cf7e86cecfbe4f7b2). Quickly testing the `hyperopt` command on an ensemble project gives very similar weights and best NDCG scores before and after this PR.

This PR improves somewhat the performance of NN ensemble suggest functionality while the results remain (only nearly?) identical (I think there were small differences in suggestion scores on my laptop). I think most of the increase comes from using `__call__` of the model instead of `predict`.

The below results are from runs at kj-kk using the current Finto AI YSO NN ensemble model (having MLLM, fastText and Omikuji base projects).
# suggest
Targeting 6 times `tests/corpora/archaeology/fulltext/*.txt`:

|                 | user time | wall time | max rss |
| --------------- | --------- | --------- | ------- |
| before (master) | 152.59  |2:35.10   | 13755484 |
| after (PR)      |  135.19  |  2:17.27 | 13757912 |

# eval
Targeting 200 documents from `kirjaesittelyt2021/yso/fin/test`:
## With 1 job
|                 | user time | wall time | max rss |F1@5 |
| --------------- | --------- | --------- | ------- |------- |
| before (master) |  174.79   |2:57.70  | 13816688 | 0.4431|
| after (PR)      | 156.12 |  2:36.77 | 13796932 |0.4431 |

## With 4 jobs
|                 | user time | wall time | max rss |F1@5 |
| --------------- | --------- | --------- | ------- |------- |
| before (master) |  188.21  | 1:45.61 | 13639148 |0.4431 |
| after (PR)      |  172.5   | 1:43.78  | 13606508 |  0.4431|